requirements.txt
altgraph==0.17.3
future==0.18.2
pefile==2022.5.30
Pillow==9.3.0
PyGetWindow==0.0.9
pyinstaller==5.6.2
pyinstaller-hooks-contrib==2022.13
PyRect==0.2.0
pywin32-ctypes==0.2.0
tinyaes==1.0.4
